solidify padding, allocate out skills component
