+++ src/pages/index.js
@@ -1,9 +1,9 @@
-import React, { useState } from "react"
+import React from "react"
 
+import IconLink from "../components/iconlink"
 import Layout from "../components/layout"
 import SEO from "../components/seo"
-import Icon from "../components/icon"
-import IconLink from "../components/iconlink"
+import Skills from "../components/skills"
 import Svg from "../components/svg"
 
 import styled from "styled-components"
@@ -17,10 +17,10 @@ const Box = styled.div`
   ${fontSize}
   ${color}
 `
+
 const Flex = styled(Box)`
   display: flex;
   flex-wrap: wrap;
-  margin-top: -1.45rem;
 `
 
 const Center = styled(Box)`
@@ -39,86 +39,12 @@ const HoverBulge = styled(Box)`
   }
 `
 
-const SkillsActivate = styled.section`
-  position: fixed;
-  cursor: pointer;
-  right: 8px;
-  bottom: 0px;
-  transition-duration: ${theme.duration.normal};
-  &.close {
-    transform: translate3d(0, 0, 0);
-  }
-  &.open {
-    transform: translate3d(150%, 0, 0);
-  }
-`
-
-const Skills = styled.section`
-  position: fixed;
-  cursor: pointer;
-  right: 8px;
-  bottom: 0px;
-`
-
-/*
-
-
-*/
-
-// transition-duration: ${theme.duration.fast};
-const SkillBox = styled(({ showSkills, ...props }) => <Box {...props} />)`
-  opacity: 0;
-  transition: transform 0.5s cubic-bezier(0.4, 0.01, 0.165, 0.99),
-    opacity 0.6s cubic-bezier(0.4, 0.01, 0.165, 0.99);
-  transition-delay: ${props => props.index * 20}ms;
-  ${({ showSkills }) =>
-    showSkills &&
-    `
-      transform: scale(1) translateY(0px);
-      opacity: 1;
-  `}
-`
-
 const Social = styled.section`
   display: flex;
   justify-content: space-evenly;
 `
 
-const skills = [
-  "illustrator",
-  "photoshop",
-  "css3",
-  "d3",
-  "docker",
-  "eslint",
-  "figma",
-  "firebase",
-  "gatsby",
-  "graphql",
-  "html5",
-  "javascript",
-  "jest",
-  "kubernetes",
-  "next",
-  "node",
-  "npm",
-  "nuxt-dot-js",
-  "postgresql",
-  "react",
-  "redux",
-  "rollup",
-  "sass",
-  "serverless",
-  "vue",
-  "yarn",
-]
-
 const IndexPage = () => {
-  const [showSkills, setShowSkills] = useState(false)
-  const _toggleShowSkills = () => {
-    setShowSkills(!showSkills)
-  }
-
   return (
     <Layout>
       <SEO
@@ -147,23 +73,7 @@ const IndexPage = () => {
         ]}
       />
 
-      <Skills
-        className={showSkills ? "open" : "close"}
-        onClick={_toggleShowSkills}
-      >
-        {skills.sort().map((skill, idx) => (
-          <SkillBox ml={2} showSkills={showSkills} index={skills.length - idx}>
-            <Icon icon={skill} inverted={false} ml={2} />
-          </SkillBox>
-        ))}
-      </Skills>
-
-      <SkillsActivate
-        className={showSkills ? "open" : "close"}
-        onClick={_toggleShowSkills}
-      >
-        <span role="img" aria-label="skills">⚡️</span>
-      </SkillsActivate>
+      <Skills />
 
       <Flex>
         <Box width={[1, 1 / 2]}>
